third_party/blink/web_tests/platform/generic/fast/forms/input-width-height-attributes-expected.txt

width and height attributes of HTMLInputElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test case #1 : Image, HTML inline setting as "160", "80"
PASS ('width' in image1) is true
PASS ('height' in image1) is true
PASS image1.width,image1.height is 160,80
Test case #2 : Image, Setting by JavaScript API as "260", "130"
PASS ('width' in image1) is true
PASS ('height' in image1) is true
PASS image2.width,image2.height is 260,130
Test case #3 : Image, Setting by JavaScript API as "120px", "60px"
PASS ('width' in image1) is true
PASS ('height' in image1) is true
PASS image3.width,image3.height is 0,0
Test case #4 : Image, Setting by JavaScript API as "120.99", "60.55"
PASS ('width' in image1) is true
PASS ('height' in image1) is true
PASS image4.width,image4.height is 120,60
Test case #5 : Text, Setting by JavaScript API as "100", "50"
PASS ('width' in text1) is true
PASS ('height' in text1) is true
PASS text1.width,text1.height is 0,0
PASS text1.width,text1.height is 100,50
PASS text1.width,text1.height is 60,40
Test case #6 : File, Setting by JavaScript API as "100", "50"
PASS ('width' in file1) is true
PASS ('height' in file1) is true
PASS file1.width,file1.height is 0,0
PASS file1.width,file1.height is 100,50
PASS file1.width,file1.height is 60,40
Test case #7 : Date, Setting by JavaScript API as "100", "50"
PASS ('width' in date1) is true
PASS ('height' in date1) is true
PASS date1.width,date1.height is 0,0
PASS date1.width,date1.height is 100,50
PASS date1.width,date1.height is 60,40
Test case #8 : Button, Setting by JavaScript API as "100", "50"
PASS ('width' in button1) is true
PASS ('height' in button1) is true
PASS button1.width,button1.height is 0,0
PASS button1.width,button1.height is 100,50
PASS button1.width,button1.height is 60,40
PASS successfullyParsed is true

TEST COMPLETE
